Refactor enemy

+++ 2DGame/02-Bubble/02-Bubble/Player.cpp
@@ -262,13 +262,13 @@ void Player::setPosition(const glm::vec2& pos)
 	sprite->setPosition(glm::vec2(float(tileMapDispl.x + posPlayer.x), float(tileMapDispl.y + posPlayer.y)));
 }
 
-glm::vec2 Player::getPosition()
+glm::ivec2 Player::getPosition()
 {
 	return posPlayer;
 }
 
 
-glm::vec2 Player::getSize()
+glm::ivec2 Player::getSize()
 {
 	return sizePlayer;
 }
